Wrap up the refactor
This last commit adds the sprite image and css creation again, cleans
up the old files (removes mostly) and updates the README a little
bit.

The output is now consistent with the new tq-db.net setup being
developed.

It's possible the loot data is still incorrect or that some other
small changes are required, but those will be implemented as the
website is updated.