Add logging for new breaking cases and fix DLC
The DLC introduced a few new errors in the file parsing that were not
being caught. In order to start indexing these errors, some basic
logging is added and each parse outputs a new log file in /logs.

This logging needs to be introduced in the already existing edge
cases too, but for now that's ignored.

This commit completes a full parse of the new DLC without errors.
Follow up commits will make the results 100% of what should be parsed
after inspection of the current results.