add heroes to boss parsing
As indicated in #14 a few boss chests were missing from the loot
tables. This led to the discovery that the missing bosses were
classified as heroes.

Heroes will need to be parsed to, doing this initially means pointing
to all creatures (and adding the classification Hero as accepted).

The follow up commits will sort out any issues discovered by
attempting to parse heroes as well.