fix missing difficulty loot and missing paths
For some reason there are a few heroes that have a placeholder '#'
when a difficulty is being skipped. Typically these would be left
empty, but apparently this edge case exists, and so needs to be
caught.

Furthermore the dbr reference shouldn't crash if the value couldn't
be parsed, so it's made more robust while testing the new hero loot
tables.