refactor: don't save image duplicates
In a past iteration, saving duplicates was necessary, but all items
stored should now be unique (with the exception of MI, which have
the same tags across difficulties).

Before doing this, the output of the sprite.css was looked into and
all existing duplicates will be removed in follow up commits.