fix: use first occurence of tags for naming
Some of the tags seem to have duplicates with completely different
names. The game chooses to use the first tag it encounters. The
parser loads all tags sequentially without checking for duplicates
so to fix this bug, simply reverse the list so the "real" first
occurence of a tag is the one we end up with.

Fixes #20