refactor: update loot chance from chest
The loot chance from a chest currently calculates the chance for each
item that comes out to be that specific item. A more realistic
indication is to multiply that chance by the number of items that
spawn from a chest.

This number is taken as the average from the max number of items that
will spawn, and the minimum number.

Closes #21