feat: add original strings to texts
Having the original strings available (without formatting) allows
some edge cases to be filled.

The edge case immediately solved is that to support locales, there
can be no hardcoded text, only references. The reference for a
100% global chance in english reads "Chance for:" or
"Chance for one of the following" in an XOR case.

This fixes that hardcoded problem.