refactor: fix skill parsing and texts
Skill parsing is nearly completed, with unique skills now being
added correctly (unique tags are now ensured). Furthermore only
skills that are granted by items or augmented by items are now added
to the output, no superfluous skills are added.

A small fix for {^...} appearing in texts is added to the Texts
class to make sure none of those appear in the output.