feat: add affix equipment types
On top of some small misc fixes the main addition is that affix
tables are now added, which in turn adds the equipment type that an
affix can be applied to.

I will put this in beta for now, as I'm not 100% sure it's fully
accurate or if I'm missing some tables, but it's definitely a good
starting point.

Closes #6