refactor: clean up parsers and run
The run.py file is cleaned up so its functions are split into
affixes, equipment, sets and skills (loot to follow).

The parsers in core.py are moved to base.py to be more consistent
with their mapping to templatebase parsers.

The locale is accepted as a parameter or otherwise defaults to en
for english, and the output file is structured tqdb.locale.version
now for its name.