refactor: assign text prefixes correctly
The Text class which takes all the text resource files and maps them
to the correct fields that are used to parse things like

offensiveFire: 2.00000 to "2 Fire Damage"

is updated to be more specific with its preparations. It also outputs
its entire table to a JSON file during parsing so it can be reviewed
while debugging and developing. The output can be found in the output
folder with the filename texts.json