test: added coverage test for base properties
The base properties that are parsed in tqdb/parsers/base.py are drawn
from the templates. To ensure that we're covering all properties in
those templates (thus guaranteeing we're parsing all properties TQ
has to offer), these tests map the properties found in templates to
the ones parsed by the parsers.