refactor: add champions to monster types
Some of the summons are marked as 'champions'. On top of this
regular champions in the game are now also added to the creature
list. The eventual monster database will hold everything so this
is a good start to continue expanding them.

This also fixes abilities not appearing on summoned pets/creatures
for scroll skills like Cyclops.