refactor: use fallback on requirements calc
It turns out quite a few items seem to not have their requirements
file set, even though in-game requirements are found.

Some digging led to the itemcost.dbr file in records/game which
seems to serve as a default (somehow...).

This commit sets it as a fallback if no static requirements were
found, and no variable ones either.