feat: add initial parsing of quest rewards
Quest rewards were tricky. It turned out the QST files can be sort of
parsed into text files where locating rewards is possible. This
allowed for assigning quests to their rewards without manually having
to go through everything.

This first iteration seems to have successfully grabbed the majority
of the quests and rewards and I'll be looking through the results
to see what I missed. Down the line things like skill points can also
be added on to this.

Refers #4