refactor: fix dynamic weight tables
Dynamic weight tables have a different meaning than what was being
interpreted by the parser. The correct implementation has two steps,
one of which has now been implemented.

The second step, which will apply the curve to items based on the
targetLevel value that the equation resolves needs some more work.

One part of it will never be fully added since it asks for the
players level. For now, we will set the average player level to
the same level as the parent (typically the monster's level).

Also, quest levels still need to figured out.

However, this commit and new output will greatly improve the
indication of an items possibility to drop at all, since it
removes items outside the level range.

Refers #36