feat: add weighted chances based on level
The targetLevelEquation is now also used to calculate the weighted
chance for items in a dynamic loot table.

The 'averagePlayerLevel' value is set to be the same as the parent
level since there is no way to input that in the parser at the
moment. Perhaps the equations will be added to the database later
to be evaluated, but the ballpark figure is good for now.

Closes #36