fix: remove affix inaccuracies
There were a few inaccuracies in the affix results. One minor error
was the incorrect listing of 'of the Mammoth' as a prefix, which is
a result of it having a prefix tag, and it appearing twice.

Furthermore, leg and legs were both used for mage/melee prefixes
as well as arm/arms. This led to half of the affix equipment types
not being counted.

The result should be a more complete and accurate index of all the
affixes and what equipment they can map to. Duplicate properties
have also been removed.

Closes #44