feat: add parsing option for all languages at once
By calling the run command with the -a or -all-languages flag the
parser will iterate over all languages instead of just the specified
or default one.

Although this takes a decent amount of time, it does simplify the
task of parsing the database in all available locales.

This is in preparation of #15