refactor: update how affixes are parsed
Aside from some black formatting and debug message fixes in this commit
the main change is around parsing affixes.

The parser now looks for the affix tables and then indexes which affix
files to actually parse.

This now includes the Tinkerer and prepares the work for the unique
affixes that only occur on some items like the Black Pearl.

Fixes #70, #60
Refers #24, #51