fix: use chance field in max iterations calc
The way the DBR files are parsed was incorrectly missing out on values
for the defensiveReflect property on the Golem's Heart charm.

There was only one value for defensiveReflect, but 5 for
defensiveReflectChance.

Now the parser checks which has more values and will use that as the
baseline for determining how many iterations there are in the DBR file.

Fixes #95